Refactor Project
new package, .gitignore to ignore files in Intellij and code in separate
classes
