+++ src/game/Corner.java
@@ -1,0 +1,11 @@
+package game;
+
+public class Corner {
+    public int x;
+    public int y;
+
+    public Corner(int x, int y) {
+        this.x = x;
+        this.y = y;
+    }
+}
