+++ src/game/Dir.java
@@ -1,0 +1,5 @@
+package game;
+
+public enum Dir {
+    left, right, up, down
+}
