+++ src/module-info.java
@@ -1,0 +1,5 @@
+module Snake.Game.in.Java.fx.by.gaspar.coding {
+    requires javafx.controls;
+    requires  javafx.fxml;
+    opens game;
+}
